requirements.txt
Flask
google-generativeai
google-api-python-client
google-auth
google_auth_oauthlib
requests
spotipy